Refactor Footer component and update import paths
- Moved Footer component to a new layout directory for better organization.
- Created a new Footer component with localization support and open source contribution links.
- Updated import statement in App.tsx to reflect the new file structure.
